[FIX] 선택 삭제 시 선택하지 않으면 삭제 불가

+++ src/components/product/ProductManageComponent.vue
@@ -113,6 +113,10 @@ const toggleAll = () => {
 }
 
 const deleteChecked = () => {
+  if (checkedProducts.value.length == 0) {
+    alert("삭제할 상품을 선택해주세요")
+    return
+  }
   if (confirm("삭제하시겠습니까?")) {
     deleteProducts(checkedProducts.value)
       .then(() => {
